pull down instanceDataSourceMap to BaseDCLIT
